requirements.txt
absl-py==2.3.1
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
pymunk==7.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==22.2
scipy==1.16.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
typing_extensions==4.14.1
Werkzeug==3.1.3
gunicorn
